Add stock data table to UI & tighten daily pull components

+++ Triowave/Models/CustomModels/StockPriceData.cs
@@ -45,6 +45,6 @@ namespace Triowave.Models.CustomModels
         public decimal Close { get; set; }
 
         [JsonPropertyName("5. volume")]
-        public long Volume { get; set; }
+        public int Volume { get; set; }
     }
 }
